refactor: 归位 RabbitMQ 消费者控制实现 --story=1070093903136621156
