+++ src/agent/aidev_agent/services/messages_handler/__init__.py
@@ -21,7 +21,6 @@ from .constants import (
 )
 from .factory import message_handler_factory
 from .in_memory import InMemoryQueueMessageHandler
-from .multi_process_mixin import MultiProcessMixin
 from .rabbitmq import RabbitMQMessageHandler
 from .single_process_mixin import SingleProcessMixin
 from .streaming_helper import GeneratorStreamingHelper
@@ -32,7 +31,6 @@ __all__ = [
     "ConsumerManagementProtocol",
     # Mixin 类
     "SingleProcessMixin",
-    "MultiProcessMixin",
     # 异常类
     "ConsumerPreemptedError",
     "RetryableHeartbeatTimeoutError",
